refactor: extract dropdown-options fetch+group into a shared helper
groupDropdownOptions/fetchDropdownOptions now live in lib/dropdown-options.ts,
used by OverpaidCases and fees-closed (previously copy-pasted fetch+group
logic in each). DropdownOptionsByCategory's canonical definition moves there
too; useDashboard.ts re-exports it for existing importers.

useDashboard.ts itself keeps its inline grouping loop rather than calling the
new helper — routing that one assignment through a function call trips a
false-positive react-hooks/set-state-in-effect on its deeply-nested async
IIFE (verified via git stash: identical behavior, only the lint result
changes). Not worth chasing for a mechanical dedup.
